Implement open streams read sharing CStructuredStorage and for archived images.  CStructuredStorage works for document files (root) storages only.  Write "_AlbumMap.txt" to image archive in Slider_v5_3.  Rename fs::CStorageTrack as fs::CStorageTrail.  Rename CThrowMode to CErrorHandler ans make it non-abstract.

+++ Slider/FileAttrAlgorithms.h
@@ -152,7 +152,7 @@ namespace func
 
 	struct AddFileSize
 	{
-		size_t operator()( size_t totalFileSize, const CFileAttr* pRight ) const
+		UINT64 operator()( UINT64 totalFileSize, const CFileAttr* pRight ) const
 		{
 			ASSERT_PTR( pRight );
 			return totalFileSize + pRight->GetFileSize();
